requirements.txt
fastapi>=0.103.1
pydantic>=2.3.0
mongoengine>=0.27.0
passlib>=1.7.4
python-multipart>=0.0.5
bcrypt>=4.0.1
uvicorn>=0.23.2
jwt>=1.3.1
python-dotenv>=1.0.0
pytest>=7.4.2